Chore: Refactor and SwitchButton
