+++ src/compnonents/competitions/EventCalendar.svelte
@@ -1,0 +1,61 @@
+<script lang="ts">
+	import type { Competition } from 'src/types/Competition';
+	import { Calendar } from 'fullcalendar';
+	import dayGridPlugin from '@fullcalendar/daygrid';
+	import { onMount } from 'svelte';
+
+	export let competitions: Competition[];
+	export let viewCalendar: boolean;
+	let calendar: Calendar | null = null;
+
+	onMount(() => {
+		let calendarEl: HTMLElement = document.getElementById('calendar')!;
+		calendar = new Calendar(calendarEl, {
+			plugins: [dayGridPlugin],
+			initialView: 'dayGridMonth',
+			displayEventTime: false,
+			eventClick: (info) => {
+				console.log(info.el);
+			},
+			eventContent: function (arg) {
+				let p1 = document.createElement('p');
+				p1.innerHTML = arg.event.title;
+				p1.style.overflow = 'clip';
+				let arrayOfDomNodes = [p1];
+				return { domNodes: arrayOfDomNodes };
+			}
+		});
+	});
+
+	function handleToggleCalendar(viewCalendar: boolean) {
+		if (!calendar) return;
+		viewCalendar ? calendar.render() : calendar.destroy();
+	}
+
+	function updateCalendarEvents(competitions: Competition[]) {
+		if (!calendar) return;
+		calendar.removeAllEvents();
+		competitions.forEach((c) => {
+			calendar?.addEvent({
+				title: c.name,
+				start: c.from,
+				end: c.to,
+				extendedProps: {
+					location: c.location,
+					nation: c.nation_code
+				}
+			});
+		});
+		if (competitions.length) calendar.gotoDate(competitions[0].from);
+	}
+
+	$: {
+		updateCalendarEvents(competitions);
+	}
+
+	$: {
+		handleToggleCalendar(viewCalendar);
+	}
+</script>
+
+<div id="calendar" />
