+++ src/routes/Rankings.svelte
@@ -2,7 +2,7 @@
 	import { Input } from 'flowbite-svelte';
 	import Line from '../compnonents/shared/Line.svelte';
 	import RankingsTable from '../compnonents/home/RankingsTable.svelte';
-	import SwitchButton from '../compnonents/home/SwitchButton.svelte';
+	import SwitchButton from '../compnonents/shared/SwitchButton.svelte';
 	import PrimaryButton from '../compnonents/shared/PrimaryButton.svelte';
 	import type { Ranking } from 'src/types/Ranking';
