+++ src/routes/competition/+page.svelte
@@ -5,9 +5,8 @@
 	import EventCard from '../../compnonents/home/EventCard.svelte';
 	import { Input } from 'flowbite-svelte';
 	import { _loadCompetitions, _loadLeagues } from './+page';
-	import { Calendar } from 'fullcalendar';
-	import dayGridPlugin from '@fullcalendar/daygrid';
-	import { onMount } from 'svelte';
+	import EventCalendar from '../../compnonents/competitions/EventCalendar.svelte';
+	import SwitchButton from '../../compnonents/shared/SwitchButton.svelte';
 
 	export let data: PageData;
 	let competitions: Competition[] = data.competitions;
@@ -15,10 +14,9 @@
 	let name: string = '';
 	let nation: string = '';
 	let leagueId: string = '';
-	let viewCalendar: boolean = false;
 	const currentDate: Date = new Date();
 	let year: number = currentDate.getFullYear();
-	let calendar: Calendar | null = null;
+	let viewCalendar: boolean = false;
 
 	async function handleSearch(year: number, name: string, nation: string, leagueId: string) {
 		competitions = await _loadCompetitions(year, name, nation, leagueId);
@@ -29,33 +27,6 @@
 		handleSearch(year, name, nation, leagueId);
 	}
 
-	$: {
-		updateCalendarEvents(competitions);
-	}
-
-	function handleToggleCalendar() {
-		if (!calendar) return;
-		viewCalendar ? calendar.destroy() : calendar.render();
-		viewCalendar = !viewCalendar;
-	}
-
-	function updateCalendarEvents(competitions: Competition[]) {
-		if (!calendar) return;
-		calendar.removeAllEvents();
-		competitions.forEach((c) => {
-			calendar?.addEvent({
-				title: c.name,
-				start: c.from,
-				end: c.to,
-				extendedProps: {
-					location: c.location,
-					nation: c.nation_code
-				}
-			});
-		});
-		if (competitions.length) calendar.gotoDate(competitions[0].from);
-	}
-
 	let timer: NodeJS.Timeout;
 	const debounce = (newValue: string) => {
 		clearTimeout(timer);
@@ -63,25 +34,6 @@
 			name = newValue;
 		}, 500);
 	};
-
-	onMount(() => {
-		let calendarEl: HTMLElement = document.getElementById('calendar')!;
-		calendar = new Calendar(calendarEl, {
-			plugins: [dayGridPlugin],
-			initialView: 'dayGridMonth',
-			displayEventTime: false,
-			eventClick: (info) => {
-				console.log(info.el);
-			},
-			eventContent: function (arg) {
-				let p1 = document.createElement('p');
-				p1.innerHTML = arg.event.title;
-				p1.style.overflow = 'clip';
-				let arrayOfDomNodes = [p1];
-				return { domNodes: arrayOfDomNodes };
-			}
-		});
-	});
 </script>
 
 <section id="competitions">
@@ -119,8 +71,12 @@
 				<option value={n.id}>{n.code}</option>
 			{/each}
 		</select>
-		<input id="toggle-calendar" type="checkbox" on:click={() => handleToggleCalendar()} />
-		<label for="toggle-calendar">View Calendar</label>
+		<SwitchButton
+			leftClickAction={() => (viewCalendar = false)}
+			rightClickAction={() => (viewCalendar = true)}
+			leftString="List"
+			rightString="Calendar"
+		/>
 	</div>
 	{#if !viewCalendar}
 		<div class="grid grid-cols-1 lg:grid-cols-2 gap-5 my-10">
@@ -129,6 +85,5 @@
 			{/each}
 		</div>
 	{/if}
-
-	<div id="calendar" />
+	<EventCalendar {competitions} {viewCalendar} />
 </section>
